% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_team_history.R
\name{plot_team_history}
\alias{plot_team_history}
\title{Plot team history at the WC}
\usage{
plot_team_history(team)
}
\arguments{
\item{team}{}
}
\value{
ggplot
}
\description{
Plot team history at the WC
}
\examples{
plot_team_history(Brazil)

}
