% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_win_probability.R
\name{plot_win_probability}
\alias{plot_win_probability}
\title{Plot win probability sequential}
\usage{
plot_win_probability(winner_seq)
}
\arguments{
\item{winners}{simulated winners as character vector}
}
\value{
ggplot
}
\description{
Plot win probability sequential
}
\examples{
top8_teams <- c("Brazil", "Germany", "France", "Netherlands", "Argentina", "Spain", "England", "Portugal")
winner_seq <- tournament_seq(teams)
plot_win_probability(winners)

}
