% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_match.R
\name{simulate_match}
\alias{simulate_match}
\title{Simulate Match at the world cup
the match is simulated based on goal scoring strength in past tournaments on the historical average scoring strength of the team}
\usage{
simulate_match(team1, team2)
}
\arguments{
\item{team1}{first team}

\item{team2}{second team}
}
\value{
simulated score and result in a data frame
}
\description{
Simulate Match at the world cup
the match is simulated based on goal scoring strength in past tournaments on the historical average scoring strength of the team
}
\examples{
simulate_match("Brazil", "Germany")
}
