% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tournament.R
\name{simulate_tournament}
\alias{simulate_tournament}
\title{Simulate the world cup knock out stages}
\usage{
simulate_tournament(teams)
}
\arguments{
\item{teams, }{character vector of 8 teams}
}
\value{
a list of all the winners and fixtures
}
\description{
Simulate the world cup knock out stages
}
\examples{
top8_teams <- c("Germany", "Brazil", "Netherlands", "France", "Argentina", "Spain", "England", "Portugal")
simulate_tournament(top8_teams)
}
