% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/team_summary.R
\name{team_summary}
\alias{team_summary}
\title{Team World Cup Summary
Performance at the previous World Cups by team}
\usage{
team_summary(team)
}
\arguments{
\item{team}{name}
}
\value{
data frame with wins/losses/draws, goals scored/conceded by team
}
\description{
Team World Cup Summary
Performance at the previous World Cups by team
}
