% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_tournament.R
\name{tournament_par}
\alias{tournament_par}
\title{Simulation Parallel Tournament}
\usage{
tournament_par(teams)
}
\arguments{
\item{teams}{character vector of 8 team names}
}
\value{
character vector of winners
}
\description{
Simulation Parallel Tournament
}
\examples{
top8_teams <- c("Brazil", "Germany", "France", "Netherlands", "Argentina", "Spain", "England", "Portugal")
tournament_par(top8_teams)
}
