% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequential_tournament.R
\name{tournament_seq}
\alias{tournament_seq}
\title{Sequential Tournament Simulation}
\usage{
tournament_seq(teams)
}
\arguments{
\item{teams}{character vector of 8 teams}
}
\value{
character vector of winners
}
\description{
Sequential Tournament Simulation
}
\examples{
top8_teams <- c("Brazil", "Germany", "France", "Netherlands", "Argentina", "Spain", "England", "Portugal")
tournament_seq(top8_teams)
}
