% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tournament.R
\name{simulate_tournament}
\alias{simulate_tournament}
\title{Simulate the world cup knock out stages}
\usage{
simulate_tournament(teams)
}
\arguments{
\item{teams}{}
}
\value{
a list of all the winners and fixtures
}
\description{
Simulate the world cup knock out stages
}
